requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Machine Learning y procesamiento de imágenes
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
scikit-image==0.21.0
pillow==10.1.0
matplotlib==3.8.2

# Utilidades
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator==2.1.0

# Logging y monitoreo
loguru==0.7.2

# Testing (opcional)
pytest==7.4.3
httpx==0.25.2